Supporting large matrices with over 2G cells

+++ src/lib/cc/basematrix.cc
@@ -9,13 +9,13 @@ namespace math {
   bool BaseMatrix::Equals(const BaseMatrix& other) const {
     if (this == &other) { return true; }
 
-    int rows = NumRows();
-    int cols = NumCols();
+    size_t rows = NumRows();
+    size_t cols = NumCols();
 
     if (rows != other.NumRows() || cols != other.NumCols()) { return false; }
 
-    for (int row = 0; row < rows; row++) {
-      for (int col = 0; col < cols; col++) {
+    for (size_t row = 0; row < rows; row++) {
+      for (size_t col = 0; col < cols; col++) {
         if ((*this)(row, col) != other(row, col)) { return false; }
       }
     }
@@ -27,43 +27,43 @@ namespace math {
     return matrix1.Equals(matrix2);
   }
 
-  void BaseMatrix::ForEach(std::function<void (int, int, float_t&)> fn) {
-    int rows = NumRows();
-    int cols = NumCols();
+  void BaseMatrix::ForEach(std::function<void (size_t, size_t, float_t&)> fn) {
+    size_t rows = NumRows();
+    size_t cols = NumCols();
 
-    for (int row = 0; row < rows; row++) {
-      for (int col = 0; col < cols; col++) {
+    for (size_t row = 0; row < rows; row++) {
+      for (size_t col = 0; col < cols; col++) {
         fn(row, col, (*this)(row, col));
       }
     }
   }
 
   void BaseMatrix::Copy(const BaseMatrix& other) {
-    int rows = std::min(NumRows(), other.NumRows());
-    int cols = std::min(NumCols(), other.NumCols());
+    size_t rows = std::min(NumRows(), other.NumRows());
+    size_t cols = std::min(NumCols(), other.NumCols());
     
-    for (int row = 0; row < rows; row++) {
-      for (int col = 0; col < cols; col++) {
+    for (size_t row = 0; row < rows; row++) {
+      for (size_t col = 0; col < cols; col++) {
         (*this)(row, col) = other(row, col);
       }
     }
   }
 
   void BaseMatrix::BufferedCopy(const BaseMatrix& other) {
-    int rows = std::min(NumRows(), other.NumRows());
-    int cols = std::min(NumCols(), other.NumCols());
+    size_t rows = std::min(NumRows(), other.NumRows());
+    size_t cols = std::min(NumCols(), other.NumCols());
     
     //TODO: See about pooling/reusing these things somehow.
     float_t * buffer = new float_t[rows * cols];
 
-    for (int row = 0; row < rows; row++) {
-      for (int col = 0; col < cols; col++) {
+    for (size_t row = 0; row < rows; row++) {
+      for (size_t col = 0; col < cols; col++) {
         buffer[row * cols + col] = other(row, col);
       }
     }
 
-    for (int row = 0; row < rows; row++) {
-      for (int col = 0; col < cols; col++) {
+    for (size_t row = 0; row < rows; row++) {
+      for (size_t col = 0; col < cols; col++) {
         (*this)(row, col) = buffer[row * cols + col];
       }
     }
@@ -72,8 +72,8 @@ namespace math {
   }
 
   std::ostream& operator<<(std::ostream& os, const BaseMatrix& matrix) {
-    for (int row = 0; row < matrix.NumRows(); row++) {
-      for (int col = 0; col < matrix.NumCols(); col++) {
+    for (size_t row = 0; row < matrix.NumRows(); row++) {
+      for (size_t col = 0; col < matrix.NumCols(); col++) {
         os << '\t' << matrix(row,col);
       }
       os << std::endl;
