+++ src/lib/h/accessor.h
@@ -1,7 +1,7 @@
 #ifndef _MDL_MATH_ACCESSOR
 #define _MDL_MATH_ACCESSOR
 
-#include "float_t.h"
+#include "typedefs.h"
 
 namespace mdl {
 namespace math {
@@ -9,14 +9,14 @@ namespace math {
 
   class DirectAccessor {
     public:
-      inline static float_t Get(float_t * data, int rowLength, int row, int col) {
+      inline static float_t Get(float_t * data, size_t rowLength, size_t row, size_t col) {
         return *(data + rowLength * row + col);
       }
-      inline static float_t& GetLRef(float_t * data, int rowLength, int row, int col) {
+      inline static float_t& GetLRef(float_t * data, size_t rowLength, size_t row, size_t col) {
         return *(data + rowLength * row + col);
       }
-      inline static int NumRows(int rows, int cols)  { return rows; }
-      inline static int NumCols(int rows, int cols)  { return cols; }
+      inline static size_t NumRows(size_t rows, size_t cols)  { return rows; }
+      inline static size_t NumCols(size_t rows, size_t cols)  { return cols; }
 
       template<typename R, typename C>
       inline static const R& GetRow(const R& row, const C& col) {
@@ -34,14 +34,14 @@ namespace math {
 
   class TransposedAccessor {
     public:
-      inline static float_t Get(float_t * data, int rowLength, int row, int col) {
+      inline static float_t Get(float_t * data, size_t rowLength, size_t row, size_t col) {
         return *(data + rowLength * col + row);
       }
-      inline static float_t& GetLRef(float_t * data, int rowLength, int row, int col) {
+      inline static float_t& GetLRef(float_t * data, size_t rowLength, size_t row, size_t col) {
         return *(data + rowLength * col + row);
       }
-      inline static int NumRows(int rows, int cols) { return cols; }
-      inline static int NumCols(int rows, int cols) { return rows; }
+      inline static size_t NumRows(size_t rows, size_t cols) { return cols; }
+      inline static size_t NumCols(size_t rows, size_t cols) { return rows; }
 
       template<typename R, typename C>
       inline static const C& GetRow(const R& row, const C& col) {
